Storage efficiency

+++ pipeline_fors2/2-sort_after_esoreflex.sh
@@ -21,6 +21,7 @@ data_dir=$(jq -r .data_dir "${param_dir}/epochs_fors2/${param_file}.json")
 data_title=${param_file}
 skip_copy=$(jq -r .skip_copy "${param_dir}/epochs_fors2/${param_file}.json")
 eso_destination=$(jq -r .esoreflex_input_dir "${config_file}")
+delete_esoreflex=$(jq -r .delete_esoreflex "${param_dir}/epochs_fors2/${param_file}.json")
 
 destination=${data_dir}/2-sorted
 mkdir "${destination}"
@@ -28,8 +29,11 @@ mkdir "${destination}"
 cd "${proj_dir}" || exit
 
 if ! ${skip_copy}; then
-  if [[ -d "${eso_destination}${data_title}" ]] ; then
-    rm -r "${eso_destination}${data_title}"
+  if ${delete_esoreflex} ; then
+    if [[ -d "${eso_destination}${data_title}" ]] ; then
+      rm -r "${eso_destination}${data_title}"
+    fi
+    python3 "${proj_dir}/pipeline_fors2/2-sort_after_esoreflex.py" --op "${data_title}" -d || exit
   fi
-  python3 "${proj_dir}/pipeline_fors2/2-sort_after_esoreflex.py" --op "${data_title}" -d || exit
+  python3 "${proj_dir}/pipeline_fors2/2-sort_after_esoreflex.py" --op "${data_title}" || exit
 fi
